requirements.txt
 fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.26.4
soundfile==0.12.1
sounddevice==0.4.6
requests==2.31.0
faster-whisper>=1.0.1
llama-cpp-python==0.3.16
pynput==1.7.6
# piper-tts==1.2.0
ddgs>=5.0.0
google-genai
python-dotenv
pytest==8.0.0
pytest-asyncio==0.23.5
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
